Day5/requirements.txt
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

numpy==1.26.4
opencv-python-headless==4.9.0.80
Pillow==10.2.0

pytesseract==0.3.10
tesseract==0.1.3
PyMuPDF==1.20.2
easyocr==1.7.1
torch==2.1.0+cpu
torchvision==0.16.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

paddlepaddle==2.6.2
paddleocr==2.7.0.3

scipy==1.11.4
deskew==1.5.1